feat: 아이콘 컴포넌트 적용 및 오류 수정
